+++ app/src/main/java/com/vecent/ssspeedtest/model/evaluter/Evaluter.java
@@ -3,6 +3,6 @@ package com.vecent.ssspeedtest.model.evaluter;
 import com.vecent.ssspeedtest.model.bean.SpeedTestResult;
 
 public interface Evaluter {
-    int evaluate(SpeedTestResult oneResult,
-                 float whiteSpeedDownLoadAvg, float blackSpeedDownLoadAvg);
+    float evaluate(SpeedTestResult oneResult,
+                   float whiteSpeedDownLoadAvg, float blackSpeedDownLoadAvg);
 }
